chore: replace old `css-to-tailwind` page and remove unnecessary dependencies
