+++ package.json
@@ -53,8 +53,7 @@
     "babel-plugin-object-styles-to-template": "^0.2.2",
     "babel-standalone": "^6.26.0",
     "clipboard-copy": "^4.0.1",
-    "css-to-tailwind": "^1.0.3",
-    "css-to-tailwindcss": "^0.2.4",
+    "css-to-tailwindcss": "^0.2.5",
     "evergreen-ui": "^4.28.0",
     "flowgen": "^1.14.1",
     "formik": "^2.2.9",
@@ -91,9 +90,6 @@
     "stringify-object": "^3.3.0",
     "svg-to-dataurl": "^1.0.0",
     "svgo": "^1.3.2",
-    "tailwindcss": "^2.2.4",
-    "tailwindcss1": "npm:tailwindcss@1.9.6",
-    "tailwindcss3": "npm:tailwindcss@3.x.x",
     "tempy": "^1.0.1",
     "text-encoding-utf-8": "^1.0.2",
     "transform-json-types": "^0.7.0",
