--- pages/api/build-tailwind-css.ts
@@ -1,12 +1,0 @@
-import { NextApiRequest, NextApiResponse } from "next";
-import tailwindCss from "@utils/tailwindcss";
-
-export default async (req: NextApiRequest, res: NextApiResponse) => {
-  try {
-    const { tailwindConfig, postCssInput } = req.body;
-    const css = await tailwindCss(tailwindConfig, postCssInput);
-    res.status(200).send(css);
-  } catch (e) {
-    res.status(500).send(e.message);
-  }
-};
